# docs/swagger.yaml
basePath: /api
definitions:
...
  ChangePasswordRequest:
    properties:
      confirmNewPassword:
        description: Must be the same as the newPassword value.
        type: string
      currentPassword:
        type: string
      newPassword:
        description: Min 6, max 150 characters.
        type: string
    required:
    - confirmNewPassword
    - currentPassword
    - newPassword
    type: object
...
  EditUser:
    properties:
      email:
        description: Must be unique
        type: string
      image:
        description: image/png or image/jpeg
        format: binary
        type: string
      username:
        description: Min 3, max 30 characters.
        type: string
    required:
    - email
    - username
    type: object
  ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  LoginRequest:
    properties:
      email:
        description: Must be unique
        type: string
      password:
        description: Min 6, max 150 characters.
        type: string
    required:
    - email
    - password
    type: object
...
  RegisterRequest:
    properties:
      email:
        description: Must be unique
        type: string
      password:
        description: Min 6, max 150 characters.
        type: string
      username:
        description: Min 3, max 30 characters.
        type: string
    required:
    - email
    - password
    - username
    type: object
  ResetPasswordRequest:
    properties:
      confirmNewPassword:
        description: Must be the same as the password value.
        type: string
      newPassword:
        description: Min 6, max 150 characters.
        type: string
      token:
        description: The from the email provided token.
        type: string
    required:
    - confirmNewPassword
    - newPassword
    - token
    type: object
  SuccessResponse:
    properties:
      success:
        description: Only returns true, not an json object
        type: boolean
    type: object
  User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      image:
        type: string
      isOnline:
        type: boolean
      updatedAt:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Valkyrie REST API Specs. This service uses sessions for authentication
  license:
    name: Apache 2.0
  title: Valkyrie API
  version: "1.0"
paths:
  /account:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
      summary: Get Current User
      tags:
      - Account
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: Update Account
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/EditUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
      summary: Update Current User
      tags:
      - Account
...
  /account/change-password:
    put:
      consumes:
      - application/json
      parameters:
      - description: Change Password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
      summary: Change Current User's Password
      tags:
      - Account
  /account/forgot-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Forgot Password
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
      summary: Forgot Password Request
      tags:
      - Account
  /account/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login account
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
      summary: User Login
      tags:
      - Account
  /account/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login account
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
      summary: User Logout
      tags:
      - Account
...
  /account/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Create account
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/User'
      summary: Create an Account
      tags:
      - Account
  /account/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reset Password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
      summary: Reset Password
      tags:
      - Account
...
swagger: "2.0"